Ispravna verzija, ispravljen id

+++ app/src/main/java/rs/aleph/android/example21/activities/SecondActivity.java
@@ -105,7 +105,7 @@ public class SecondActivity extends AppCompatActivity
         navigationView.setNavigationItemSelectedListener(this);
 
 
-        id = getIntent().getExtras().getLong(Home.ATRAKCIJA);
+        id = getIntent().getExtras().getInt(Home.ATRAKCIJA);
 
 
         try {
